fear: initial release

+++ src/bot/interactions/delete.interaction.ts
@@ -1,16 +1,25 @@
 import { Injectable } from '@nestjs/common';
 import { ButtonInteraction } from 'discord.js';
+import { CheckRightsService } from '../checkRights.service';
+import { MESSAGES } from 'src/common/constants';
 
 @Injectable()
 export class DeleteChannelInteraction {
+  constructor(private readonly checkRightsService: CheckRightsService) {}
   async onButtonInteract(interaction: ButtonInteraction) {
     await interaction.deferReply({ flags: 'Ephemeral' });
+
+    const { isOwner } = await this.checkRightsService.check(interaction);
+
+    if (!isOwner) {
+      await interaction.editReply(MESSAGES.NO_RIGHTS);
+      return;
+    }
+
     const member = interaction.guild.members.cache.get(interaction.user.id);
     const channel = member.voice.channel;
-    //TODO: Требуется интеграция с БД для проверки на владельца канала
+
     await channel.delete();
-    await interaction.editReply({
-      content: `Голосовой канал **${channel.name}** удален.`,
-    });
+    await interaction.editReply(`Голосовой канал **${channel.name}** удален.`);
   }
 }
